opencode/enterprise-kb/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
pypdf==5.1.0
markdown==3.7
langfuse==2.59.0
openai==1.58.1
structlog==24.4.0
ragas==0.2.12
datasets==3.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
aiosqlite==0.20.0
